Revert "perf: 修改部门接口"
This reverts commit 8e66082aab0d2523a4612d9d706619428e48a1c6.

+++ src/main/java/com/example/workflow/model/dto/DeptFormDto.java
@@ -13,7 +13,7 @@ public class DeptFormDto {
     /**
      * 父部门等级
      */
-    private Short parentLevel;
+    private Integer parentLevel;
     /**
      * 部门名称
      */
